Added Foreign Keys

+++ db/migrate/20190611094151_create_diagnosis.rb
@@ -1,7 +1,7 @@
 class CreateDiagnosis < ActiveRecord::Migration[5.2]
   def change
     create_table :diagnosis, :primary_key => :diagnosis_id do |t|
-    	t.bigint    :encounter_id
+    	t.integer    :encounter_id
       t.integer    :concept_id
     	t.boolean    :primary_diagnosis
       t.boolean    :secondary_diagnosis
