+++ db/migrate/20190611094731_create_breastfeeding_statuses.rb
@@ -1,7 +1,7 @@
 class CreateBreastfeedingStatuses < ActiveRecord::Migration[5.2]
   def change
     create_table    :breastfeeding_statuses, :primary_key => :breastfeeding_status_id do |t|
-      t.bigint       :encounter_id
+      t.integer       :encounter_id
       t.integer     :concept_id
       t.integer     :value_coded
       t.boolean     :voided, null: false, default: 0
